examples/demo_app/app.py
#!/usr/bin/env python
# -*- coding: utf-8 -*-

"""
Demo application for the NVIDIA GenAI Retail Accelerator.

This Streamlit app demonstrates the capabilities of the multi-modal
retail GenAI system, allowing users to:
1. Upload product images for recognition
2. Ask natural language questions about products
3. View recommendations and insights
"""

import os
import sys
import numpy as np
import pandas as pd
import torch
import streamlit as st
from pathlib import Path
from PIL import Image
import time

# Add parent directory to path for imports
sys.path.append(str(Path(__file__).parent.parent.parent))

# Import project modules
# In a real implementation, these would be properly installed packages
# from src.models import load_model
# from src.inference import predict

# Placeholder functions for demo
def load_model(model_path, device="cuda"):
    """Placeholder for loading model."""
    time.sleep(1)  # Simulate loading time
    return "model_loaded"

def process_image(image, model):
    """Placeholder for processing images."""
    time.sleep(0.5)  # Simulate processing time
    return {
        "product_id": np.random.randint(1, 101),
        "confidence": np.random.uniform(0.8, 0.99),
        "bbox": [10, 10, 100, 100]
    }

def answer_question(question, context, model):
    """Placeholder for question answering."""
    time.sleep(0.5)  # Simulate processing time
    
    # Simple rule-based responses for demo
    if "price" in question.lower():
        return f"The price of this product is ${np.random.uniform(10, 100):.2f}."
    elif "available" in question.lower() or "in stock" in question.lower():
        return "Yes, this product is currently in stock at your nearest store."
    elif "similar" in question.lower() or "recommend" in question.lower():
        return "Based on this product, I would recommend Product 42, Product 78, and Product 23."
    else:
        return "This product is made with high-quality materials and is one of our bestsellers."


def main():
    """Main function for the demo application."""
    st.set_page_config(
        page_title="Retail GenAI Demo",
        page_icon="🛒",
        layout="wide"
    )
    
    st.title("🛒 NVIDIA GenAI Retail Accelerator")
    st.subheader("Multi-Modal AI for Retail Applications")
    
    # Set up sidebar
    st.sidebar.image("https://developer.nvidia.com/sites/default/files/akamai/nvidia-logo.png", width=200)
    st.sidebar.title("Controls")
    
    # Model selection
    model_type = st.sidebar.selectbox(
        "Select Model Type",
        ["Full Multi-Modal", "Vision Only", "Language Only"]
    )
    
    # Load model when user clicks button
    if st.sidebar.button("Load Model"):
        with st.spinner("Loading model... This might take a few seconds."):
            model = load_model("path/to/model")
            st.sidebar.success("Model loaded successfully!")
            st.session_state.model = model
    else:
        st.sidebar.warning("Click 'Load Model' to initialize the system")
        if "model" not in st.session_state:
            st.session_state.model = None
    
    # Main content area - split into two columns
    col1, col2 = st.columns([3, 2])
    
    with col1:
        st.header("Product Recognition & Analysis")
        
        # Image upload
        uploaded_file = st.file_uploader("Upload a product or shelf image", type=["jpg", "jpeg", "png"])
        
        if uploaded_file is not None:
            # Display the uploaded image
            image = Image.open(uploaded_file)
            st.image(image, caption="Uploaded Image", use_column_width=True)
            
            # Process the image if model is loaded
            if st.session_state.model is not None:
                with st.spinner("Analyzing image..."):
                    result = process_image(image, st.session_state.model)
                
                # Display results
                st.success(f"Product identified with {result['confidence']:.2%} confidence!")
                
                # Display product details
                st.subheader("Product Details")
                st.json({
                    "product_id": result["product_id"],
                    "name": f"Product {result['product_id']}",
                    "category": np.random.choice(["Electronics", "Clothing", "Groceries", "Home"]),
                    "price": f"${np.random.uniform(10, 100):.2f}",
                    "in_stock": np.random.choice([True, False], p=[0.8, 0.2])
                })
    
    with col2:
        st.header("Ask about this product")
        
        # Only enable question asking if an image has been uploaded and processed
        if uploaded_file is not None and st.session_state.model is not None:
            question = st.text_input("What would you like to know about this product?")
            
            if question:
                with st.spinner("Thinking..."):
                    answer = answer_question(question, {"product_id": result["product_id"]}, st.session_state.model)
                
                st.markdown(f"**Answer:** {answer}")
                
                # Display thinking process for educational purposes
                with st.expander("See AI reasoning"):
                    st.write("""
                    1. Analyzed product features from the image
                    2. Retrieved product details from catalog
                    3. Understood question intent: price inquiry
                    4. Generated natural language response
                    """)
            
            # Show recommendations
            st.subheader("You might also like")
            recommendation_cols = st.columns(3)
            
            for i, col in enumerate(recommendation_cols):
                with col:
                    st.image("https://via.placeholder.com/100", caption=f"Product {np.random.randint(1, 101)}")
        else:
            st.info("Upload an image and load the model to ask questions about products")
    
    # Performance metrics
    st.header("NVIDIA GPU Performance")
    perf_cols = st.columns(3)
    
    with perf_cols[0]:
        st.metric("Inference Speed", "12 ms/image", delta="-85% vs CPU")
    
    with perf_cols[1]:
        st.metric("GPU Utilization", "72%")
    
    with perf_cols[2]:
        st.metric("Throughput", "83 images/sec", delta="+14x vs CPU")


if __name__ == "__main__":
    main()
